requirements.txt
Flask
flask_session
pyrebase4
mediapipe
opencv-python
numpy
pyttsx3
gtts
pydub
pygame
requests
matplotlib
